split conftest file
