+++ app/database/database.py
@@ -4,7 +4,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
 
-SQLALCHEMY_DATABASE_URL = os.getenv("DATABASE_CONNECTION_STRING")
+SQLALCHEMY_DATABASE_URL = os.getenv("DATABASE_CONNECTION_STRING", "sqlite://")
 
 engine = create_engine(
     SQLALCHEMY_DATABASE_URL, connect_args={"check_same_thread": False}
